Fix nasty bug
I already knew about it but it was fixed only in parallel branch.

+++ include/qx/Utils.hpp
@@ -13,7 +13,7 @@ static inline bool getBit(std::size_t x, std::size_t index) {
 }
 
 static inline void setBit(std::size_t &x, std::size_t index, bool value) {
-    x = (x & ~(static_cast<size_t>(1) << index)) | (value << index);
+    x = (x & ~(static_cast<std::size_t>(1) << index)) | (static_cast<std::size_t>(value) << index);
 }
 
 // std::bitset is slightly slower than this.
